+++ src/api/observable.ts
@@ -95,6 +95,7 @@ function createObservable(v: any, arg2?: any, arg3?: any) {
 
 export interface IObservableFactory {
     // observable overloads
+    (value: number | string | null | undefined | boolean): never // Nope, not supported, use box
     (target: Object, key: string, baseDescriptor?: PropertyDescriptor): any // decorator
     <T>(value: T[], options?: CreateObservableOptions): IObservableArray<T>
     <K, V>(value: Map<K, V>, options?: CreateObservableOptions): ObservableMap<K, V>