+++ CHANGELOG.md
@@ -115,6 +115,7 @@ All the random notes that should make up a nice changelog:
 * `autorun`, `reaction` and `when` now can take an `onError` property as parameter
 * `toJS` now accepts options: `{    detectCycles?: boolean    exportMapsAsObjects?: boolean }`, both `true` by default
 * `when` now supports an `timeout` option, so `mobxUtils.whenWithTimeout` is no longer needed
+* when `when` is invoked withoout effect, it will automatically return a promise, so that you can await conditions. These promises are canceallable by calling `.cancel`
 
 ## Breaking changes